Add support for branch prediction and bh
The branch prediction feature available for ppc64 architecture,
gives a hint to the processor if a conditional branch is likely or
unlikely to be done. In order to make this work, it is necessary to
set the "bo" field for a branch instruction. The first two "bo" bits
designate the likelihood of a branch (see the BO enum class).

This commit also supports the "bh" field. This feature is used to
give some hints if the bctr is part of a loop. Currently, this
feature cannot be used in hhvm.