Port tc-print functions to PPC64
The tc-print only uses two functions with arch code specific, which were
ported to PPC64 using the internal Decoder.
This implementation is still not good to send to facebook, since the target
of the branchs are not being dispayed and we need to think in a better
architectural solution, since it's using static "if" to switch the code.